fix(gates): report mypy crashes as could-not-run; unblock refit on version drift
Barnacles #3, #5, #6 from the rulebook-ai onboarding:

#5 (blocker) — overconfidence:missing-annotations.py reported
"FAILED: 0 type error(s) found" with empty output whenever mypy exited
non-zero for a NON-type reason (here: duplicate module names from vendored
template copies). The user got a red gate with nothing to fix. Now: when mypy
exits non-zero but yields zero parseable type errors, it failed to RUN — the
gate returns a could-not-run ERROR carrying mypy's own message ("Duplicate
module named 'llm_api'") plus its suggested resolutions. Same three-state
contract missing tools already use. Genuine type errors still FAIL.

#3 — sm_env.tool_inventory FAILed on version drift, which hard-blocked
`sm refit --start` for anyone one patch release behind (black 26.5.0 vs a
26.5.1 pin). A tool that is present but stale still runs, so drift now WARNs;
missing tools and validator rejects still FAIL (they genuinely can't run).

#6 — myopia:ambiguity-mines.py supports exclude_dirs but its fix text never
said so, unlike its sibling repeated-code. Vendored copies and test helpers
should be excluded, not refactored; the hint now says that and prints it.

Tests: mypy crash → ERROR with mypy's words; real type errors → FAIL;
version drift → WARN while validator rejects stay FAIL. The static-analysis
tests moved to their own file — slop-mop's own code-sprawl gate blocked the
commit at 1002/1000 lines, which is the tool working.
